\name{hbrwted1-package}
\alias{hbrwted1-package}
\alias{hbrwted1}
\docType{package}
\title{
\packageTitle{hbrwted1}
}
\description{
\packageDescription{hbrwted1}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{hbrwted1}
\packageIndices{hbrwted1}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{hbrwted1}

Maintainer: \packageMaintainer{hbrwted1}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
